requirements.txt
langchain
langchain-openai
fastmcp
fastapi
uvicorn
python-dotenv
sqlalchemy
openai
